+++ mods/sbz_logic/link_tool.lua
@@ -36,10 +36,16 @@ local function render_individual_link(pos, name)
     end
 end
 
-local function render_links()
+local timer = 0
+local function render_links(dtime)
+    timer = timer + dtime
+    if timer < render_links_delay then return end
+    timer = 0
+
     for k, v in pairs(waypoint_ids) do
         sbz_api.remove_waypoint(v)
     end
+    waypoint_ids = {}
     for k, v in pairs(minetest.get_connected_players()) do
         local wielded_item = v:get_wielded_item()
         if wielded_item:get_name() == "sbz_logic:luacontroller_linker" then
@@ -67,10 +73,9 @@ local function render_links()
             end
         end
     end
-    minetest.after(render_links_delay, render_links)
 end
 
-minetest.after(0, render_links) -- me elegant!
+core.register_globalstep(render_links)
 
 local function try_to_link_to_luac(stack, pos, placer)
     local meta = stack:get_meta()